Convert some components to TS
